sync part1-helloworld branch with docs
Reviewed in https://codereview.appspot.com/9727046/
