+++ app.yaml
@@ -6,12 +6,4 @@ threadsafe: true
 
 handlers:
 - url: /.*
-  script: guestbook.application
-
-
-libraries:
-- name: webapp2
-  version: latest
-- name: jinja2
-  version: latest
-
+  script: helloworld.application
